src/test/java/com/blogspot/ostas/leetcode/all/medium/all_paths_from_source_to_target/Solution.java
package com.blogspot.ostas.leetcode.all.medium.all_paths_from_source_to_target;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;

/*
    	797. all paths from source to target

https://leetcode.com/problems/all-paths-from-source-to-target

Given a directed acyclic graph (DAG) of n nodes labeled from 0 to n - 1, find all possible paths from node 0 to node n - 1 and return them in any order.
The graph is given as follows: graph[i] is a list of all nodes you can visit from node i (i.e., there is a directed edge from node i to node graph[i][j]).

*/
class Solution {
    public List<List<Integer>> allPathsSourceTarget(int[][] graph) {
        List<List<Integer>> ans = new ArrayList<>();
        dfs(graph, 0, new ArrayList<>(Arrays.asList(0)), ans);
        return ans;
    }

    private void dfs(int[][] graph, int u, List<Integer> path, List<List<Integer>> ans) {
        if (u == graph.length - 1) {
            ans.add(new ArrayList<>(path));
            return;
        }

        for (final int v : graph[u]) {
            path.add(v);
            dfs(graph, v, path, ans);
            path.remove(path.size() - 1);
        }
    }
}
